requirements.txt
flask
pandas
python-dotenv
openai
langchain
langchain-community
langchain-mistralai
chromadb
scikit-learn